Refactor storage helper test classes
